+++ projects/rocprim/CHANGELOG.md
@@ -8,6 +8,7 @@ Full documentation for rocPRIM is available at [https://rocm.docs.amd.com/projec
 
 * Added support for large number of segments and large number of items to `rocprim::device_segmented_radix_sort`.
 * Added fixed-size-segments variant for `rocprim::device_segmented_reduce`.
+* Added support for large number of segments to `rocprim::device_segmented_reduce`.
 
 ## rocPRIM 5.0.0 for ROCm 8.0